+++ src/DragAndDrop/Draggable/components/DragLayer.js
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import {Portal} from 'react-portal';
 import PropTypes from 'prop-types';
 import {DragLayer} from 'react-dnd';
 
@@ -30,35 +31,55 @@ const onOffsetChange = monitor => {
 };
 
 class CustomDragLayer extends React.Component {
+  shouldRenderLayer = (props = this.props) => {
+    const {id, item, itemType, draggedType, isDragging} = props;
+    return isDragging && id === item.id && itemType === draggedType;
+  };
+
+  renderPreview = () => {
+    const {offsetOfHandle} = this.props;
+    const styles = Object.assign({}, layerStyles, {left: -offsetOfHandle.x, top: -offsetOfHandle.y});
+    return (
+      <div
+        style={styles}
+        ref={node => dragPreviewRef = node}
+        >
+        {this.props.renderPreview({})}
+      </div>
+    );
+  }
+
+  renderPreviewInPortal = () => {
+    return (
+      <Portal>
+        {this.renderPreview()}
+      </Portal>
+    );
+  }
+
   render() {
-    const {
-      item,
-      itemType,
-      draggedType,
-      isDragging,
-      renderPreview,
-      id
-    } = this.props;
-    const shouldRenderLayer = isDragging && id === item.id && itemType === draggedType;
-    if (!shouldRenderLayer) {
+    if (!this.shouldRenderLayer()) {
       return null;
     }
-    return <div style={layerStyles} ref={node => dragPreviewRef = node}>{renderPreview({})}</div>;
+    return this.props.usePortal ? this.renderPreviewInPortal() : this.renderPreview();
   }
 }
 
 CustomDragLayer.propTypes = {
+  offsetOfHandle: PropTypes.object,
   item: PropTypes.object,
   itemType: PropTypes.string,
   draggedType: PropTypes.string,
   isDragging: PropTypes.bool,
   renderPreview: PropTypes.func,
+  usePortal: PropTypes.bool,
   id: PropTypes.oneOfType([PropTypes.string, PropTypes.number])
 };
 
-export default DragLayer(monitor => {
+export default DragLayer((monitor, props) => {
   onOffsetChange(monitor);
   return {
+    offsetOfHandle: props.offsetOfHandle,
     item: monitor.getItem(),
     itemType: monitor.getItemType(),
     isDragging: monitor.isDragging()