+++ src/SideMenu/core/SideMenu.js
@@ -3,14 +3,14 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import styles from './styles.scss';
 
-const SideMenu = ({children, inFlex, className}) => {
+const SideMenu = ({children, inFlex, className, dataHook}) => {
   const rootStyles = classNames({
     [styles.root]: true,
     [styles.inFlex]: inFlex
   }, className);
 
   return (
-    <div className={rootStyles} data-hook="side-menu">
+    <div className={rootStyles} data-hook={dataHook}>
       {children}
     </div>
   );
@@ -23,7 +23,8 @@ SideMenu.defaultProps = {
 SideMenu.propTypes = {
   inFlex: PropTypes.bool,
   className: PropTypes.string,
-  children: PropTypes.node
+  children: PropTypes.node,
+  dataHook: PropTypes.string
 };
 
 export default SideMenu;