+++ stories/Tooltip/Composite/ExamplePopoverEmptyState.js
@@ -6,18 +6,17 @@ import EmptyState from 'wix-style-react/EmptyState';
 
 const ExamplePopoverEmptyState = () => (
   <Tooltip
+    popover
     content={(
       <EmptyState
         title="You don't have labels yet"
         subtitle="Once you create some labels you'll see a list in here"
         />
     )}
-    hideTrigger="click"
     placement="right"
     padding={24}
     minWidth={300}
     shouldCloseOnClickOutside
-    showTrigger="click"
     dataHook="popover-empty-state"
     >
     <Button>