+++ stories/Tooltip/Composite/Template.js
@@ -8,7 +8,9 @@ import TooltipForEyesOnly from './TooltipForEyesOnly';
 export class Template extends Component {
 
   static propTypes = {
+    padding: PropTypes.any,
     onChange: PropTypes.func.isRequired,
+    popover: PropTypes.bool,
     theme: Tooltip.propTypes.theme,
     placement: Tooltip.propTypes.placement,
     tooltipContent: Tooltip.propTypes.content,
@@ -42,6 +44,8 @@ export class Template extends Component {
   getComponent() {
     return (
       <Tooltip
+        padding={this.props.padding}
+        popover={this.props.popover || this.props.type === 'popover'}
         placement={this.props.placement}
         alignment="center"
         content={this.props.tooltipContent}