+++ testkit/polyfills/request-animation-frame.js
@@ -1,16 +1,21 @@
+const uninstall = () => {
+  delete global.requestAnimationFrame;
+  delete global.cancelAnimationFrame;
+};
+
 const install = () => {
   let lastTime = 0;
   const vendors = ['ms', 'moz', 'webkit', 'o'];
 
-  for (let x = 0; x < vendors.length && !global.requestAnimationFrame; ++x) {
-    global.requestAnimationFrame = global[vendors[x] + 'RequestAnimationFrame'];
+  for (let x = 0; x < vendors.length && !window.requestAnimationFrame; ++x) {
+    window.requestAnimationFrame = window[vendors[x] + 'RequestAnimationFrame'];
     window.cancelAnimationFrame =
       window[vendors[x] + 'CancelAnimationFrame'] ||
       window[vendors[x] + 'CancelRequestAnimationFrame'];
   }
 
-  if (!global.requestAnimationFrame) {
-    global.requestAnimationFrame = function (callback) {
+  if (!window.requestAnimationFrame) {
+    window.requestAnimationFrame = function (callback) {
       const currTime = new Date().getTime();
       const timeToCall = Math.max(0, 16 - (currTime - lastTime));
       const id = window.setTimeout(() => {
@@ -26,6 +31,9 @@ const install = () => {
       clearTimeout(id);
     };
   }
+
+  global.requestAnimationFrame = global.requestAnimationFrame || window.requestAnimationFrame;
+  global.cancelAnimationFrame = global.cancelAnimationFrame || window.cancelAnimationFrame;
 };
 
-export default {install};
+export default {install, uninstall};